Cython wrapper
