+++ .gitignore
@@ -1,3 +1,6 @@
+deps/
+wrapper.c
+
 build_tmp/
 .idea
 *.dylib
