+++ hygese/__init__.py
@@ -1,1 +1,1 @@
-from .hygese import *
+from hygese.solver import AlgorithmParameters, Solver
