+++ pyproject.toml
@@ -1,3 +1,2 @@
 [build-system]
-requires = ["setuptools"]
-build-backend = "setuptools.build_meta"
+requires = ["setuptools", "numpy", "Cython"]
