+++ setup.py
@@ -1,6 +1,10 @@
 from setuptools import setup, find_packages
 from setuptools.command.build_ext import build_ext as _build_ext
 from setuptools.command.build_py import build_py as _build_py
+
+from distutils.core import setup, Extension
+from Cython.Build import cythonize
+
 import subprocess
 import os
 import platform
@@ -10,13 +14,6 @@ import shutil
 import urllib.request
 
 urlretrieve = urllib.request.urlretrieve
-
-# try:
-#     import urllib.request
-#     urlretrieve = urllib.request.urlretrieve
-# except ImportError:  # python 2
-#     from urllib import urlretrieve
-
 
 # read the contents of your README file
 from pathlib import Path
@@ -40,9 +37,9 @@ def _safe_makedirs(*paths):
 HGS_VERSION = "2.0.0"
 HGS_SRC = f"https://github.com/vidalt/HGS-CVRP/archive/v{HGS_VERSION}.tar.gz"
 
-LIB_DIR = "lib"
-BUILD_DIR = "lib/build"
-BIN_DIR = "lib/bin"
+LIB_DIR = "deps"
+BUILD_DIR = "deps/build"
+BIN_DIR = "deps/bin"
 
 
 def get_lib_filename():
@@ -76,37 +73,29 @@ def download_build_hgs():
     shutil.copyfile(f"{BUILD_DIR}/{LIB_FILENAME}", f"hygese/{LIB_FILENAME}")
 
 
-# LIB_VERSION = "0.0.1"
-# HGS_CVRP_WIN = f"https://github.com/chkwon/Libhgscvrp_jll.jl/releases/download/libhgscvrp-v{LIB_VERSION}%2B0/" + \
-#                 f"libhgscvrp.v{LIB_VERSION}.x86_64-w64-mingw32-cxx11.tar.gz"
-
-# def download_binary_hgs():
-#     print(HGS_CVRP_WIN)
-
-#     _safe_makedirs(LIB_DIR)
-#     dll_tarball_name = "win_bin.tar.gz"
-#     hgs_bin_path = pjoin(LIB_DIR, dll_tarball_name)
-#     urlretrieve(HGS_CVRP_WIN, hgs_bin_path)
-#     _run(f"tar xzvf {dll_tarball_name}", LIB_DIR)
-#     shutil.copyfile(f"{BIN_DIR}/{LIB_FILENAME}", f"hygese/{LIB_FILENAME}")
-
-
 class BuildPyCommand(_build_py):
     def run(self):
-        print("Build!!!!!! Run!!!!")
-
-        if platform.system() == "Windows":
-            # download_binary_hgs()
-            download_build_hgs()
-        else:
-            download_build_hgs()
-
+        download_build_hgs()
         _build_py.run(self)
+
+
+
+extentions = [
+    Extension(
+        name="hygese.wrapper",
+        sources=["hygese/wrapper.pyx"],
+        include_dirs = [f"deps/HGS-CVRP-{HGS_VERSION}/Program/"],
+        library_dirs = ["hygese"],
+        libraries = ["hgscvrp"],
+    )
+]
+
+
 
 
 setup(
     name="hygese",
-    version="0.0.0.8",
+    version="0.0.1.0",
     description="A Python wrapper for the HGS-CVRP solver",
     long_description=long_description,
     long_description_content_type="text/markdown",
@@ -127,6 +116,7 @@ setup(
     cmdclass={
         "build_py": BuildPyCommand,
     },
+    ext_modules=cythonize(extentions),    
     package_data={
         "": ["libhgscvrp.*"],
     },
